+++ event-sourcing/subscription-manager/src/test/java/uk/gov/justice/services/event/sourcing/subscription/catchup/consumer/manager/ConcurrentEventStreamConsumerManagerTest.java
@@ -37,9 +37,6 @@ public class ConcurrentEventStreamConsumerManagerTest {
     @Mock
     private ConsumeEventQueueTaskManager consumeEventQueueTaskManager;
 
-    @Mock
-    private EventQueueConsumerFactory eventQueueConsumerFactory;
-
     @Spy
     private EventStreamsInProgressList eventStreamsInProgressList = new EventStreamsInProgressList();
 
@@ -63,12 +60,11 @@ public class ConcurrentEventStreamConsumerManagerTest {
 
         when(eventsInProcessCounterProvider.getInstance()).thenReturn(eventsInProcessCounter);
         when(eventsInProcessCounter.maxNumberOfEventsInProcess()).thenReturn(false);
-        when(eventQueueConsumerFactory.create(concurrentEventStreamConsumerManager)).thenReturn(eventQueueConsumer);
         when(publishedEvent.getStreamId()).thenReturn(streamId);
 
         concurrentEventStreamConsumerManager.add(publishedEvent, subscriptionName, catchupCommand, commandId);
 
-        verify(consumeEventQueueTaskManager).consume(eventQueueCaptor.capture(), eq(eventQueueConsumer), eq(subscriptionName), eq(catchupCommand), eq(commandId));
+        verify(consumeEventQueueTaskManager).consume(eventQueueCaptor.capture(), eq(subscriptionName), eq(catchupCommand), eq(commandId));
 
         final Queue<PublishedEvent> events = eventQueueCaptor.getValue();
         assertThat(events.size(), is(1));
@@ -92,14 +88,13 @@ public class ConcurrentEventStreamConsumerManagerTest {
 
         when(eventsInProcessCounterProvider.getInstance()).thenReturn(eventsInProcessCounter);
         when(eventsInProcessCounter.maxNumberOfEventsInProcess()).thenReturn(false);
-        when(eventQueueConsumerFactory.create(concurrentEventStreamConsumerManager)).thenReturn(eventQueueConsumer);
         when(publishedEvent_1.getStreamId()).thenReturn(streamId);
         when(publishedEvent_2.getStreamId()).thenReturn(streamId);
 
         concurrentEventStreamConsumerManager.add(publishedEvent_1, subscriptionName, catchupCommand, commandId);
         concurrentEventStreamConsumerManager.add(publishedEvent_2, subscriptionName, catchupCommand, commandId);
 
-        verify(consumeEventQueueTaskManager).consume(eventQueueCaptor.capture(), eq(eventQueueConsumer), eq(subscriptionName), eq(catchupCommand), eq(commandId));
+        verify(consumeEventQueueTaskManager).consume(eventQueueCaptor.capture(), eq(subscriptionName), eq(catchupCommand), eq(commandId));
 
         final Queue<PublishedEvent> eventsStream = eventQueueCaptor.getValue();
         assertThat(eventsStream.size(), is(2));
@@ -125,14 +120,13 @@ public class ConcurrentEventStreamConsumerManagerTest {
 
         when(eventsInProcessCounterProvider.getInstance()).thenReturn(eventsInProcessCounter);
         when(eventsInProcessCounter.maxNumberOfEventsInProcess()).thenReturn(false);
-        when(eventQueueConsumerFactory.create(concurrentEventStreamConsumerManager)).thenReturn(eventQueueConsumer);
         when(publishedEvent_1.getStreamId()).thenReturn(streamId_1);
         when(publishedEvent_2.getStreamId()).thenReturn(streamId_2);
 
         concurrentEventStreamConsumerManager.add(publishedEvent_1, subscriptionName, catchupCommand, commandId);
         concurrentEventStreamConsumerManager.add(publishedEvent_2, subscriptionName, catchupCommand, commandId);
 
-        verify(consumeEventQueueTaskManager, times(2)).consume(eventQueueCaptor.capture(), eq(eventQueueConsumer), eq(subscriptionName), eq(catchupCommand), eq(commandId));
+        verify(consumeEventQueueTaskManager, times(2)).consume(eventQueueCaptor.capture(), eq(subscriptionName), eq(catchupCommand), eq(commandId));
 
         final List<Queue<PublishedEvent>> allValues = eventQueueCaptor.getAllValues();
 
@@ -163,13 +157,12 @@ public class ConcurrentEventStreamConsumerManagerTest {
 
         when(eventsInProcessCounterProvider.getInstance()).thenReturn(eventsInProcessCounter);
         when(eventsInProcessCounter.maxNumberOfEventsInProcess()).thenReturn(false);
-        when(eventQueueConsumerFactory.create(concurrentEventStreamConsumerManager)).thenReturn(eventQueueConsumer);
         when(publishedEvent_1.getStreamId()).thenReturn(streamId_1);
         when(publishedEvent_2.getStreamId()).thenReturn(streamId_2);
 
         concurrentEventStreamConsumerManager.add(publishedEvent_1, subscriptionName, catchupCommand, commandId);
 
-        verify(consumeEventQueueTaskManager).consume(eventQueueCaptor.capture(), eq(eventQueueConsumer), eq(subscriptionName), eq(catchupCommand), eq(commandId));
+        verify(consumeEventQueueTaskManager).consume(eventQueueCaptor.capture(), eq(subscriptionName), eq(catchupCommand), eq(commandId));
 
         final Queue<PublishedEvent> eventsStream_1 = eventQueueCaptor.getValue();
         assertThat(eventsStream_1.size(), is(1));
@@ -178,7 +171,7 @@ public class ConcurrentEventStreamConsumerManagerTest {
         concurrentEventStreamConsumerManager.isEventConsumptionComplete(new FinishedProcessingMessage(eventsStream_1));
         concurrentEventStreamConsumerManager.add(publishedEvent_2, subscriptionName, catchupCommand, commandId);
 
-        verify(consumeEventQueueTaskManager, times(2)).consume(eventQueueCaptor.capture(), eq(eventQueueConsumer), eq(subscriptionName), eq(catchupCommand), eq(commandId));
+        verify(consumeEventQueueTaskManager, times(2)).consume(eventQueueCaptor.capture(), eq(subscriptionName), eq(catchupCommand), eq(commandId));
 
         final Queue<PublishedEvent> eventsStream_2 = eventQueueCaptor.getValue();
         assertThat(eventsStream_2.size(), is(1));
@@ -205,5 +198,18 @@ public class ConcurrentEventStreamConsumerManagerTest {
         concurrentEventStreamConsumerManager.decrementEventsInProcessCount();
 
         verify(eventsInProcessCounter).decrementEventsInProcessCount();
+    }
+
+    @Test
+    public void shouldDecrementTheEventsInProcessCountByGivenNumber() throws Exception {
+
+        final int count = 2;
+        final EventsInProcessCounter eventsInProcessCounter = mock(EventsInProcessCounter.class);
+
+        when(eventsInProcessCounterProvider.getInstance()).thenReturn(eventsInProcessCounter);
+
+        concurrentEventStreamConsumerManager.decrementEventsInProcessCountBy(count);
+
+        verify(eventsInProcessCounter).decrementEventsInProcessCountBy(count);
     }
 }